assimp: work around gcc bug on SuperH
gcc versions earlier than gcc 6.x fail to build assimp on SuperH when
static linking:

  AssxmlExporter.cpp:623:1: error: unable to find a register to spill in class 'GENERAL_REGS'

It's the combination of -Os and *not* having -fPIC that makes gcc
fail, which explains why configurations with dynamic linking work
fine.

 -Os -fPIC -> works
 -Os       -> fails
 -O2 -fPIC -> works
 -O2       -> works

Therefore, as a workaround, we are forcing the use of -O2 on SuperH
when the gcc version is older than gcc 6.x and we're statically
linking.

Fixes:

  http://autobuild.buildroot.net/results/ec88aa8118179e30e24603cc45292047dca19216/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>