pacakge/dieharder: fix autoreconf
When dieharder was committed, Thomas removed the hook to cleanup the m4
files, on the reason that autoreconf would recreate the broken symlinks
(and because the hook was too complex).

It turns out the hook was needed: if the symbolic links are
broken (libtool not installed on the host machine), autoreconf fails to
do its job.

Reinstate a simpler hook.

Fixes:
    http://autobuild.buildroot.org/results/a92/a92e9a74b0966f8f0bc4080f918d5ddf4c1ec0ce/
    http://autobuild.buildroot.org/results/a4f/a4f6647f69ad89ff1ffe5f1331281fb7a57ade4e/

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Julien Viard de Galbert <julien@vdg.name>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>