triggerhappy: bumped to 0.5.0 to fix systemd unit
This bumps the version of triggerhappy to 0.5.0
The systemd unit supplied by buildroot assumed that filename
globbing was supported by thd. But this was just recently
added. So with this version bump usage of "/dev/input/event*"
is now possible. Systemd Unit is adjusted accordingly.

eventtable.h patch was removed as this was patched upstream.

This patch fixes bug #9836.

Signed-off-by: Michael Heinemann <posted@heine.so>