fio: re-enable on uClibc
The fio package was disabled on uClibc because posix_madvise() was
back then not implemented in an official release of uClibc. It now
exists in uClibc-ng since its version 1.0.0, so we can assume it's
available. We already make lots of assumptions about external uClibc
toolchains, and they should definitely be using a recent version of
uClibc-ng to be usable by Buildroot.

Also, the Config.in comment was not handling this uClibc exception.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>